Revert "Change: components/Shared -> components/Web, components/Game으로 이동"
