+++ package.json
@@ -18,9 +18,9 @@
     "react-dom": "^18.2.0",
     "react-responsive": "^9.0.2",
     "react-router-dom": "^6.17.0",
-    "react-scripts": "^5.0.1",
+    "react-scripts": "5.0.1",
     "recoil": "^0.7.7",
-    "styled-components": "^5.3.11",
+    "styled-components": "^6.1.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
