--- src/components/Web/Shared/Form/PasswordRequirementStyle.jsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const PasswordRequirementWrapper = styled.div``;
-
-export const PasswordRequirementMeassage = styled.p`
-  font-size: ${({ theme }) => theme.fontSize.xxxs};
-`;
