+++ src/pages/Web/Notice/NoticeList.jsx
@@ -1,26 +1,7 @@
-import React, { useState, useEffect } from "react";
-
-import { ContentWrapper, WideContent } from "@/styles/CommonStyle";
-import Header from "@/components/Web/Shared/Layout/Header";
-import Pagination from "@/components/Web/Shared/Board/Pagination";
+import React from "react";
 
 const NoticeList = () => {
-  const [currPage, setCurrPage] = useState(16);
-
-  useEffect(() => {}, [currPage]);
-
-  return (
-    <ContentWrapper row="center" col="center">
-      <WideContent dir="col">
-        <Header type="detail" />
-        <Pagination
-          currPage={currPage}
-          setCurrPage={setCurrPage}
-          totalPageIdx="19"
-        ></Pagination>
-      </WideContent>
-    </ContentWrapper>
-  );
+  return <h1>NoticeList Page</h1>;
 };
 
 export default NoticeList;
